When resetting streams, delete our state immediately.
This has a subtle effect, but it's important: we want all actions on
streams that *we* reset to be handled as connection-level errors, rather
than stream-level ones. We want it to be as though that stream no longer
exists, so that the connection discovers the problem rather than the
stream state machine. This avoids the need to communicate to the streams
that they have been reset so that they handle errors differently.

The main reason we do this is to avoid overcommitting space. While it's
logically more sensible to push this knowledge down into the stream (it
already knows it!), if we didn't have any logic in the connection then
we'd need to keep reset streams around forever in order to allow their
state machines to handle the inputs. That's potentially a very large
state commmitment because H2Stream objects and their associated entities
are really quite large. On the other hand, handling this in the
connection requires a much smaller state commitment: a set of integers
and then a few extra code paths.

For that reason, we prefer to have the connection handle this mess
*always*, rather than either a hybrid solution where both the connection
and stream need to know (thus duplicating code) or where the stream
always knows but we have a nasty state commitment to handle the case.

(cherry picked from commit 872012ff017ce1b070fcfc0604ffd92c92932eec)