Handle pushes on closed streams.
Pushes on closed streams are tricky.

If we know we reset the stream ourselves then we should assume that the
push was made in good faith and simply reset the stream it opened.
However, if the stream was closed in any other way (because of a remote
RST_STREAM or the natural course of the state machine) then this is a
severe violation of the protocol. We can't just issue a basic RST_STREAM
for it (which is what we've been doing) because there's now a *new*
stream that the remote peer believes exists: the one that was just
pushed. We could RST that as well, but at a certain point you just need
to give up on a peer that is misbehaving so badly and issue a
GOAWAY(PROTOCOL_ERROR). That's what we do here.

(cherry picked from commit bc056cfc2b4261219277aee7232413bf8fb2cbd8)