Handle headers as bytes internally.
This changes our internal representation of headers to use bytes
everywhere, rather than unicode.

hyper-h2 actually previously played a bit fast-and-loose with header
format internally, meaning that on Python 2 there was quite a bit of
implicit encoding/decoding going on when we worked with headers. This
should remove that: headers are bytes up until right when we're done
with them, at which point they *may* become unicode.

Note that becoming unicode is now optional.