Update H2Configuration to take logger param
H2Configuration now takes a logger parameter which allows us to pass in
a logger that confirms to the requirements of this module, so that debug
logging can be performed on H2Connection and H2Stream.  A DummyLogger
class has also been added which is used as the default if no
logger is passed in to avoid cluttering up code with conditionals.